refactor(core-storage-api): consolidate router validation guards into _validation
The fail-closed ``_require(body, key)`` guard was byte-identical across four
routers (evolve, skill_factory, insights, entities), and ``entities`` had added
``_require_number`` while ``evolve`` still hand-rolled the same numeric check
inline. Extract both into ``core_storage_api/routers/_validation.py`` and import
them across the four routers; replace evolve's inline ``isinstance(..., (int,
float))`` loop with ``_require_number``.

Pure refactor, no behavior change to the happy path or the 422 detail messages.
One intentional tightening: evolve's delta/floor/cap now go through
``_require_number``, which rejects ``bool`` (a subclass of ``int`` that was
never a valid tuning value) — the 422 message is unchanged. Surfaced during the
Fix 2 Ph6 (#477) /simplify review.
