+++ core-storage-api/src/core_storage_api/routers/insights.py
@@ -21,17 +21,11 @@ from datetime import datetime
 
 from fastapi import APIRouter, HTTPException, Request
 
+from core_storage_api.routers._validation import _require
 from core_storage_api.services.postgres_service import PostgresService
 
 router = APIRouter(tags=["Insights"])
 _svc = PostgresService()
-
-
-def _require(body: dict, key: str) -> str:
-    val = body.get(key)
-    if not val:
-        raise HTTPException(status_code=422, detail=f"{key} is required")
-    return val
 
 
 def _scope_args(body: dict) -> tuple[str, str | None, str, str]:
